Gets everything but internal_api_server compiling.

+++ src/pfs/route/sharder.go
@@ -8,22 +8,22 @@ import (
 )
 
 type sharder struct {
-	numShards   int
-	numReplicas int
+	numShards   uint64
+	numReplicas uint64
 }
 
-func newSharder(numShards int, numReplicas int) *sharder {
+func newSharder(numShards uint64, numReplicas uint64) *sharder {
 	return &sharder{numShards, numReplicas}
 }
 
-func (s *sharder) NumShards() int {
+func (s *sharder) NumShards() uint64 {
 	return s.numShards
 }
 
-func (s *sharder) NumReplicas() int {
+func (s *sharder) NumReplicas() uint64 {
 	return s.numReplicas
 }
 
-func (s *sharder) GetShard(file *pfs.File) (int, error) {
-	return int(adler32.Checksum([]byte(path.Clean(file.Path)))) % s.numShards, nil
+func (s *sharder) GetShard(file *pfs.File) (uint64, error) {
+	return uint64(adler32.Checksum([]byte(path.Clean(file.Path)))) % s.numShards, nil
 }
